+++ src/reducers/auth.js
@@ -1,7 +1,6 @@
 import {
     LOGIN_REQUEST,
     LOGIN_SUCCESS,
-    LOGIN_FAILURE,
     LOGOUT_SUCCESS,
     AUTH_INIT_ERROR,
     AUTH_INIT_SUCCESS,
@@ -9,8 +8,10 @@ import {
     PASSWORD_RESET_EMAIL_SUCCESS,
     RESET_REQUEST,
     RESET_SUCCESS,
-    AUTH_FAILURE
-} from '../actions/user';
+    AUTH_FAILURE,
+    REGISTER_REQUEST,
+    REGISTER_SUCCESS
+} from '../actions/auth';
 
 const token = localStorage.getItem('token');
 export default function auth(state = {
@@ -19,41 +20,34 @@ export default function auth(state = {
     errorMessage: '',
     currentUser: null,
     loadingInit: true
-}, action) {
-    switch (action.type) {
-        case RESET_REQUEST:
+}, {type, payload}) {
+    switch (type) {
         case LOGIN_REQUEST:
+        case RESET_REQUEST:
+        case PASSWORD_RESET_EMAIL_REQUEST:
+        case REGISTER_REQUEST:
             return Object.assign({}, state, {
                 isFetching: true,
-                isAuthenticated: false,
+                errorMessage: '',
+                //isAuthenticated: false,
             });
-        case PASSWORD_RESET_EMAIL_REQUEST:
         case LOGIN_SUCCESS:
+        case LOGOUT_SUCCESS:
+        case RESET_SUCCESS:
+        case PASSWORD_RESET_EMAIL_SUCCESS:
+        case REGISTER_SUCCESS:
             return Object.assign({}, state, {
                 isFetching: false,
-                isAuthenticated: true,
                 errorMessage: '',
             });
         case AUTH_FAILURE:
             return Object.assign({}, state, {
                 isFetching: false,
-                errorMessage: action.payload,
-            });
-        case RESET_SUCCESS:
-        case LOGIN_FAILURE:
-            return Object.assign({}, state, {
-                isFetching: false,
-                isAuthenticated: false,
-                errorMessage: action.payload,
-            });
-        case PASSWORD_RESET_EMAIL_SUCCESS:
-        case LOGOUT_SUCCESS:
-            return Object.assign({}, state, {
-                isAuthenticated: false,
+                errorMessage: payload,
             });
         case AUTH_INIT_SUCCESS:
             return Object.assign({}, state, {
-                currentUser: action.payload.currentUser || null,
+                currentUser: payload.currentUser || null,
                 loadingInit: false
             });
         case AUTH_INIT_ERROR: